メディア，オブジェクトにおける座標系は整数にした． また，キュービックの生成が正しく行われていなかったので修正した．

+++ CGH-ver2.0.0/CGH-ver2.0.0/object.h
@@ -11,14 +11,14 @@ public:
 	void P();
 	void P_lineX(double size, int resolution);
 	void cubic(int size, int resolution,int distance);
-	void initial_position(double x, double y, double z, vector<vector<double>> object_position);
-	vector<vector<double>> point;
+	void initial_position(double x, double y, double z, vector<vector<int>> object_position);
+	vector<vector<int>> point;
 };
 
 class Media {
 public:
 	void media_criate(int mediasize_X, int mediasize_Y, int pixcel_pitch);
 
-	vector<vector<vector<double>>> point;
-	void initial_position(double x, double y, double z, vector<vector<vector<double>>> point_inf);
+	vector<vector<vector<int>>> point;
+	void initial_position(double x, double y, double z, vector<vector<vector<int>>> point_inf);
 };
